make the examples compile again :)
