+++ .gitignore
@@ -11,3 +11,4 @@ haddock_badge.json
 haddocks/
 *.prof
 *.prof.html
+hgeometry-examples/pub/*.wasm
