+++ cabal.project
@@ -15,7 +15,7 @@ packages:
   hgeometry-svg
   hgeometry-geojson
   hgeometry
-  -- hgeometry-examples
+  hgeometry-examples
   -- ../hiraffe
   -- ../miso-bulma
   -- ../approximate-equality
